Making secure and login one file as auth.go
